seperated Animal and NeedBar
