+++ student-projects/gilad/vlad-notebook/sessions-12-19-typescript/session-18/home-project/app.ts
@@ -1,293 +1,7 @@
 import { Diets, AnimalDiet, foodList } from "./AnimalDiet.js";
 import { Gender, AnimalType, camel, bear } from "./AnimalType.js";
-class NeedBar {
-	private name: string;
-	private minValue: number;
-	private maxValue: number;
-	private value: number;
-	private alertValue: number;
-	private listenerAlertReached: Function;
-
-	constructor(
-		name: string,
-		minValue: number,
-		maxValue: number,
-		startValue?: number,
-		alertValue?: number,
-		listenerAlertValueReached?: Function
-	) {
-		this.name = name;
-		this.minValue = minValue;
-		this.maxValue = maxValue;
-		this.alertValue = alertValue ? alertValue : 0;
-		this.setValue(startValue ? startValue : 0);
-		this.setAlertValueListener(listenerAlertValueReached);
-	}
-
-	private normalizeBarValue(amount: number) {
-		amount = Math.floor(amount);
-
-		amount =
-			amount > this.maxValue
-				? (amount = this.maxValue)
-				: amount < this.minValue
-				? (amount = this.minValue)
-				: (amount = amount);
-
-		return amount;
-	}
-
-	setValue(amount: number) {
-		this.value = this.normalizeBarValue(amount);
-		if (this.value <= this.alertValue) this.callAlertListener();
-	}
-
-	setFull() {
-		this.setValue(100);
-	}
-
-	setEmpty() {
-		this.setValue(0);
-	}
-
-	addValue(amount: number) {
-		this.setValue(this.getValue() + amount);
-	}
-
-	reduceValue(amount: number) {
-		this.setValue(this.getValue() - amount);
-	}
-
-	setAlertValueListener(listener: Function | undefined) {
-		if (listener) this.listenerAlertReached = listener;
-	}
-
-	callAlertListener() {
-		this.listenerAlertReached(this.value);
-	}
-
-	getName() {
-		return this.name + " bar";
-	}
-
-	getValue() {
-		return this.value;
-	}
-
-	getAlertValue() {
-		return this.alertValue;
-	}
-
-	getStatus() {
-		return `${this.getName()} is currently ${this.getValue} out of ${
-			this.maxValue
-		}`;
-	}
-
-	isAlertingValue() {
-		return this.getValue() <= this.getAlertValue() ? true : false;
-	}
-}
-
-class Animal {
-	private birthId: number;
-	private name: string;
-	private gender: Gender;
-	private type: AnimalType;
-	private hunger: NeedBar;
-	private hydration: NeedBar;
-	private happiness: NeedBar;
-
-	constructor(id: number, name: string, gender: Gender, type: AnimalType) {
-		this.birthId = id;
-		this.name = name;
-		this.gender = gender;
-		this.type = type;
-
-		this.hunger = new NeedBar("hunger", 0, 100, 100, 25, this.hungryAlert);
-		this.hydration = new NeedBar(
-			"hydration",
-			0,
-			100,
-			100,
-			50,
-			this.thirstyAlert
-		);
-		this.happiness = new NeedBar("happiness", 0, 4, 4, 2, this.sadAlert);
-	}
-
-	getBirthID() {
-		return this.birthId;
-	}
-
-	getName() {
-		return this.name;
-	}
-
-	getGender() {
-		return this.gender;
-	}
-
-	getType() {
-		return this.type;
-	}
-
-	getHungerBar() {
-		return this.hunger;
-	}
-
-	getHydrationBar() {
-		return this.hydration;
-	}
-
-	getHappinessBar() {
-		return this.happiness;
-	}
-
-	getSound() {
-		return this.getType().getSound();
-	}
-
-	isHungry() {
-		return this.getHungerBar().isAlertingValue();
-	}
-
-	isThirsty() {
-		return this.getHydrationBar().isAlertingValue();
-	}
-
-	isSad() {
-		return this.getHappinessBar().isAlertingValue();
-	}
-
-	isHappy() {
-		return !this.isSad();
-	}
-
-	eat(food: Food, amount?: number) {
-		if (!this.getType().getDiet().canEat(food)) {
-			this.announce(
-				`can't eat ${food.name}, ${
-					this.getGender() === Gender.Female ? "s" : ""
-				}he's a ${this.getType().getName()}!`
-			);
-		} else {
-			if (amount) {
-				this.getHungerBar().addValue(amount);
-				this.announce(` eaten ${amount} ${food.name}!`);
-			} else {
-				this.getHungerBar().setFull();
-				this.announce(` eaten ${food.name}!`);
-			}
-		}
-	}
-
-	drink(amount?: number) {
-		if (amount) {
-			this.getHydrationBar().addValue(amount);
-			this.announce(`drinked ${amount} water!`);
-		} else {
-			this.getHydrationBar().setFull();
-			this.announce(`drinked until hydration!`);
-		}
-	}
-
-	addHappiness(amount?: number) {
-		if (amount) {
-			this.getHappinessBar().addValue(amount);
-			this.announce(`is now happier by ${amount}!`);
-		} else {
-			this.getHappinessBar().setFull();
-			this.announce(`is now fully happy!`);
-		}
-	}
-
-	reduceHappiness(amount?: number) {
-		if (amount) {
-			this.getHappinessBar().reduceValue(amount);
-			this.announce(`is now sadder by ${amount}!`);
-		} else {
-			this.getHappinessBar().setEmpty();
-			this.announce(`is now fully sad!`);
-		}
-	}
-
-	announce(message: string) {
-		console.log(`${this.toString()} ${message} ${this.getSound()}`);
-	}
-
-	toString() {
-		return `${this.getGender()} ${this.getType().getName()} ${this.getType().getEmoji()} - ${this.getName()}`;
-	}
-
-	hungryAlert() {
-		console.log(`${this.type.getName()} - ${this.name} is hungry!`);
-	}
-
-	thirstyAlert() {
-		console.log(`${this.type.getName()} - ${this.name} is thirsty!`);
-	}
-
-	sadAlert() {
-		console.log(`${this.type.getName()} - ${this.name} is sad!`);
-	}
-}
-
-class AnimalList {
-	private name: string;
-	private animals: Animal[];
-
-	constructor() {
-		this.animals = [];
-	}
-
-	add(animal: Animal) {
-		if (animal) this.animals.push(animal);
-		console.log(`${animal.toString()} added to the list`);
-	}
-
-	getAll() {
-		return this.animals;
-	}
-
-	getAnimalsByName(name: string) {
-		const animalsByName = this.animals.filter(
-			(animal) => animal.getName() === name
-		);
-
-		return animalsByName;
-	}
-
-	getAnimalsByType(type: AnimalType) {
-		return this.animals.filter((animal) => animal.getType() === type);
-	}
-
-	getAnimalsByTypeName(typeName: string) {
-		return this.animals.filter(
-			(animal) => animal.getType().getName() === typeName
-		);
-	}
-
-	getAnimalsByHungry() {
-		return this.animals.filter((animal) => animal.isHungry());
-	}
-
-	getAnimalsByThirsty() {
-		return this.animals.filter((animal) => animal.isThirsty());
-	}
-
-	getAnimalsBySad() {
-		return this.animals.filter((animal) => animal.isSad());
-	}
-
-	getAnimalsByHappy() {
-		return this.animals.filter((animal) => animal.isHappy());
-	}
-
-	getAnimalByID(id: number) {
-		return this.animals.find((animal) => animal.getBirthID());
-	}
-}
+import { Animal, AnimalList } from "./Animal.js";
+import { NeedBar } from "./NeedBar.js";
 
 enum Biome {
 	grassland,
